+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of the nginx cookbook.
 ## Unreleased
 
 - Enable unified_mode resource for Chef 17
+- Remove dependency on the deprecated ohai cookbook
 
 ## 11.5.3 - *2021-04-20*